Added fast enum library.
