+++ PhotoViewer/PhotoViewer.csproj
@@ -34,6 +34,7 @@
   </ItemGroup>
 
   <ItemGroup>
+    <PackageReference Include="FastEnum" Version="1.7.0" />
     <PackageReference Include="MaterialDesignThemes.MahApps" Version="0.2.0" />
     <PackageReference Include="MetadataExtractor" Version="2.7.1" />
     <PackageReference Include="Prism.Wpf" Version="8.1.97" />
